U.S.#3 descriptive handlers
